requirements.txt
pandas
openpyxl
numpy
google-generativeai
langchain-core
gunicorn
fastapi
uvicorn
pydantic

